Major rework of the conversion loop
Now we have on the lowest level the application name, which the COM factory will try to load.
This is the first step of automatically loading each possible SolidWorks version.
The benefit is, that if one is not bought (so far) or not correctly uninstalled, that the algorithm will fallback to the next and retry again.
However, if the the file was made with eg. 2017 and 2016 is the only working, it won't help of course..