SolidWorksReader: Sorting and reversing versions
So the most recent version is used before any other.
The sense is clear: A file has been created with SldWks 2018, but the plugin will try to use older versions first. However, 2018 files are not compatible with older versions. This is also the case for any other version.
To prevent a waste of time, we being with the most recent one.